Inductor changes for fake tensor
[ghstack-poisoned]
