+++ torchdynamo/optimizations/training.py
@@ -329,13 +329,15 @@ class CudaGraphModule(Module):
 
 
 def find_input_mutations(g):
-    FK = "fake_result"
+    def meta_fk(meta):
+        return meta["val"] if "val" in meta else meta["fake_result"]
+
     inputs = defaultdict(set)
     input_idx = 0
     mutated_inputs = set()
     for n in g.nodes:
         if n.op == "placeholder":
-            inputs[StorageWeakRef(n.meta[FK].storage())].add(input_idx)
+            inputs[StorageWeakRef(meta_fk(n.meta).storage())].add(input_idx)
             input_idx += 1
         elif n.op == "call_function":
             if n.target is operator.getitem:
@@ -356,7 +358,7 @@ def find_input_mutations(g):
                     # TODO: not correct for args that contain tensors in a struct
                     # like list
                     mutated_inputs |= inputs[
-                        StorageWeakRef(argument.meta[FK].storage())
+                        StorageWeakRef(meta_fk(argument.meta).storage())
                     ]
         # TODO: error on unrecognized nodes
     return mutated_inputs
